pyproject.toml
[project]
name = "rubricas-qdrant"
version = "0.1.0"
description = "Sistema de generación de rúbricas con Qdrant Vector Search y LangSmith"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "google-adk>=0.1.0",
    "sentence-transformers>=2.2.0",
    "qdrant-client>=1.7.0",
    "langsmith>=0.4.26",
    "python-dotenv>=1.0.0",
    "pypdf>=4.0.0",
    "fastapi>=0.123.10",
    "uvicorn>=0.40.0",
    "python-multipart>=0.0.22",
    "langchain>=1.2.9",
    "langchain-google-genai>=4.2.0",
    "a2a-sdk[http-server]>=0.3.2",
    "click>=8.0.0",
    "starlette>=0.47.2",
    "httpx>=0.27.0",
    "beeai-framework>=0.1.77",
    "langgraph>=0.0.10",
]
